+++ src/archive.rs
@@ -1,3 +1,5 @@
+//! Older versions of the algorithms; kept around for development and testing
+
 pub mod biotsavart;
 pub mod dualtree;
 pub mod octree;