+++ src/oersted/constants.py
@@ -1,3 +1,5 @@
 from numpy import pi
 
+""" Magnetic permeability of free space
+"""
 MU0: float = 4.0 * pi * 10**-7