+++ src/vec3.rs
@@ -1,4 +1,5 @@
-/// A 3-length vector
+//! A 3-length vector
+
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 #[repr(transparent)]
 pub struct Vec3(pub [f64; 3]);