+++ tests/test_solenoid.py
@@ -12,12 +12,22 @@ Note: there's some sort of units mismatch with gmsh
 
 import oersted
 from oersted import Mesh, DirectSolver, OctreeSolver
+from oersted.testing import bz_finite_length_solenoid
 import numpy as np
 
+# ---
+# Solenoid properties
+# ---
+ri: float = 0.025  # (m)
+ro: float = 0.050  # (m)
+dr: float = ro - ri
+length: float = 0.250  # (m)
+jmag: float = 1e8  # (A/m2)
 
-def check_solenoid():
+
+def test_solenoid():
+
     mesh_size: float = 15.0
-    jmag: float = 1e8
     theta: float = 0.1
     nthreads: int = 0
     ntargets_axis: int = 100
@@ -35,12 +45,11 @@ def check_solenoid():
     jdensity[:, 1] = jmag * np.cos(phi)
 
     # ---
-    # Solution at center of solenoid
+    # Solution on axis of solenoid
     # ---
 
     targets_axis = np.zeros((ntargets_axis, 3))
     targets_axis[:, 2] = np.linspace(-0.125, 0.125, ntargets_axis)
-    # targets_axis[:,0] = np.linspace(0, 0.10, ntargets_axis)
 
     bdirect_pt_axis = oersted.b_field(mesh.to_centroid_mesh(), jdensity, targets_axis, solver=direct_solver)
     boctree_pt_axis = oersted.b_field(mesh.to_centroid_mesh(), jdensity, targets_axis, solver=octree_solver)
@@ -61,9 +70,27 @@ def check_solenoid():
     assert err_octree_pt_axis < 1e-2
     assert err_octree_tet_axis < 1e-2
 
-    #
+    # ---
+    # Check solution against analytical at centroid of solenoid
+    # ---
+
+    bz_analytical: float = oersted.MU0 * jmag * dr
+    r_avg = 0.5 * (ro + ri)
+    target = np.array([[0.0, 0.0, 0.0]])
+    bz_analytical: float = bz_finite_length_solenoid(jmag, length, r_avg, dr, 0.0)
+    bz_tet4_direct = oersted.b_field(mesh, jdensity, target, solver=direct_solver)[0, 2]
+    bz_tet4_octree = oersted.b_field(mesh, jdensity, target, solver=octree_solver)[0, 2]
+    bz_point_direct = oersted.b_field(mesh, jdensity, target, solver=direct_solver)[0, 2]
+    bz_point_octree = oersted.b_field(mesh, jdensity, target, solver=octree_solver)[0, 2]
+
+    assert np.abs(bz_analytical - bz_tet4_direct) / bz_analytical < 1e-2
+    assert np.abs(bz_analytical - bz_tet4_octree) / bz_analytical < 1e-2
+    assert np.abs(bz_analytical - bz_point_direct) / bz_analytical < 1e-2
+    assert np.abs(bz_analytical - bz_point_octree) / bz_analytical < 1e-2
+
+    # ---
     # Solve for self-fields
-    #
+    # ---
 
     targets = mesh.centroids
 
@@ -85,10 +112,6 @@ def check_solenoid():
     assert err_mesh_pt_octree < 1e-1  # pt method known to be inaccurate inside the mesh
     assert err_mesh_pt_direct < 1e-1
     assert err_mesh_tet_octree < 1e-2
-
-
-def test_solenoid():
-    check_solenoid()
 
 
 if __name__ == "__main__":